requirements.txt
appdirs==1.4.3
packaging==16.8
pycryptodome==3.4.6
pyparsing==2.2.0
six==1.10.0
